Having something that looks like a source mapping in your source file causes the plugin to not work.
According to the [spec](https://sourcemaps.info/spec.html):

"The generated code may include a line at the end of the source, with the following form: "

So changing the logic to only allow the last source mapping should do the trick.

Maybe in the future all matches should be checked until a working one is found. Starting with last one.